[SPARK-46433][PS][TESTS][FOLLOWUPS] Reorganize `OpsOnDiffFramesGroupByTests`: Factor out remaining slow tests
### What changes were proposed in this pull request?
Factor out remaining slow tests in `OpsOnDiffFramesGroupByTests`

### Why are the changes needed?
to make it more suitable for parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44418 from zhengruifeng/ps_test_diff_groupby_3.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
