+++ python/pyspark/pandas/tests/connect/diff_frames_ops/test_parity_groupby_transform.py
@@ -16,19 +16,21 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_ops_on_diff_frames_groupby import OpsOnDiffFramesGroupByTestsMixin
+from pyspark.pandas.tests.diff_frames_ops.test_groupby_transform import GroupByTransformMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 
 
-class OpsOnDiffFramesGroupByParityTests(
-    OpsOnDiffFramesGroupByTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
+class GroupByTransformParityTests(
+    GroupByTransformMixin,
+    PandasOnSparkTestUtils,
+    ReusedConnectTestCase,
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_groupby import *  # noqa: F401
+    from pyspark.pandas.tests.connect.diff_frames_ops.test_parity_groupby_transform import *  # noqa
 
     try:
         import xmlrunner  # type: ignore[import]
